Moved Slurm tests originally written by Matea to the slurm_test.py file to reflect module layout
